binance-to-influxdb/requirements.txt
influxdb-client
requests
websockets
pyyaml
